test: fix order_manager test failures (reduce from 39 to 33)
- Fixed test_bracket_orders.py (8/12 tests passing)
  - Created proper test implementation with both mixins
  - Fixed BracketOrderResponse attribute access
  - Added missing attributes (stats, position_manager)

- Fixed ALL test_position_orders.py tests (18/18 passing)
  - Changed position mocks to use 'type' instead of 'side'
  - Aligned with PositionType enum values

- Fixed test_utils.py alignment test expectation
- Fixed test_tracking.py async decorator issues
- Fixed test_error_recovery.py OrderPlaceResponse initialization

Progress: 179 passing (+6), 33 failing (-6)
